Groq Speaks
These are much smaller, cheaper models than Claude haiku & sonnet 3.5

They are not as bright, but that's OK, they're cheap! We can use them a
lot longer without hitting our budget limits. I don't know how much this
costs, but it should be pennies compared to - we've already spent $4 on
Claude Code this month and we're not trying to spend that much - we just
need to operate mecris (and document our experiments/experiences)

Signed-off-by: Kingdon B <kingdon@urmanac.com>