tell Groq to create a plan for the agent
We're going to install the tamagotchi, give the agent some instructions
(mostly what it says in this file) and then tell him to run continuously
until finished or blocked. And the tamagotchi is going to keep him on
task, I guess?

Groq experiment has been a success even if we don't close the loop here.

Signed-off-by: Kingdon B <kingdon@urmanac.com>